+++ src/Sulu/Bundle/ContentBundle/Tests/Functional/Controller/NodeControllerTest.php
@@ -1978,6 +1978,52 @@ class NodeControllerTest extends SuluTestCase
         $this->assertEquals('Dornbirn', $result['title']);
     }
 
+    public function testRenamePageWithLinkedChild()
+    {
+        $client = $this->createAuthenticatedClient();
+        $this->importer->import(__DIR__ . '/../../app/Resources/exports/tree.xml');
+
+        $document = $this->documentManager->find('585ccd35-a98e-4e41-a62c-e502ca905496', 'en');
+        $document->setStructureType('internallinks');
+        $document->getStructure()->bind(
+            [
+                'internalLinks' => [
+                    '5778b19f-460a-47fc-93da-9a6126e5c384',
+                ],
+            ]
+        );
+        $this->documentManager->persist($document, 'en');
+        $this->documentManager->publish($document, 'en');
+        $this->documentManager->flush();
+        $this->documentManager->clear();
+
+        $client->request('GET', '/api/nodes/' . $document->getUuid() . '?webspace=sulu_io&language=en');
+        $data = json_decode($client->getResponse()->getContent(), true);
+        $data['title'] = 'Sulu is awesome';
+
+        $client->request(
+            'PUT',
+            '/api/nodes/' . $document->getUuid() . '?webspace=sulu_io&language=en&action=publish',
+            $data
+        );
+
+        $this->assertHttpStatusCode(200, $client->getResponse());
+        $data = json_decode($client->getResponse()->getContent(), true);
+
+        $this->assertEquals('/sulu-is-awesome', $data['path']);
+
+        /** @var SessionInterface $liveSession */
+        $liveSession = $this->getContainer()->get('sulu_document_manager.live_session');
+        /** @var SessionInterface $session */
+        $session = $this->getContainer()->get('sulu_document_manager.default_session');
+
+        $node = $liveSession->getNode('/cmf/sulu_io/contents/sulu-is-awesome');
+        $this->assertEquals($data['id'], $node->getIdentifier());
+
+        $node = $session->getNode('/cmf/sulu_io/contents/sulu-is-awesome');
+        $this->assertEquals($data['id'], $node->getIdentifier());
+    }
+
     private function setUpContent($data)
     {
         $client = $this->createAuthenticatedClient();