+++ src/Sulu/Bundle/SnippetBundle/Controller/SnippetController.php
@@ -18,6 +18,7 @@ use FOS\RestBundle\Routing\ClassResourceInterface;
 use FOS\RestBundle\View\View;
 use FOS\RestBundle\View\ViewHandler;
 use PHPCR\NodeInterface;
+use Sulu\Bundle\SnippetBundle\Document\SnippetDocument;
 use Sulu\Bundle\SnippetBundle\Snippet\DefaultSnippetManagerInterface;
 use Sulu\Bundle\SnippetBundle\Snippet\SnippetRepository;
 use Sulu\Component\Content\Compat\StructureManagerInterface;
@@ -235,6 +236,17 @@ class SnippetController implements SecuredControllerInterface, ClassResourceInte
         $document = $this->findDocument($uuid, $this->getLocale($request));
 
         $this->requestHashChecker->checkHash($request, $document, $document->getUuid());
+        if (!$this->checkAreaSnippet($request, $document)) {
+            return new JsonResponse(
+                [
+                    'structures' => [],
+                    'other' => [],
+                    'isDefault' => true,
+                ],
+                409
+            );
+        }
+
         $this->processForm($request, $document);
 
         return $this->handleView($document);
@@ -545,5 +557,16 @@ class SnippetController implements SecuredControllerInterface, ClassResourceInte
         $view = View::create($document)->setContext($context);
 
         return $this->viewHandler->handle($view);
+    }
+
+    private function checkAreaSnippet(Request $request, SnippetDocument $document)
+    {
+        $force = $request->headers->get('SuluForcePut', false);
+        $structureType = $request->request->get('template');
+
+        return $force
+            || $structureType === $document->getStructureType()
+            || !$this->defaultSnippetManager->isDefault($document->getUuid())
+            || $structureType === $this->defaultSnippetManager->loadType($document->getUuid());
     }
 }