+++ src/Sulu/Component/CustomUrl/Tests/Unit/Manager/Manager/CustomUrlManagerTest.php
@@ -24,62 +24,110 @@ use Sulu\Component\DocumentManager\Metadata;
 use Sulu\Component\DocumentManager\MetadataFactoryInterface;
 use Sulu\Component\DocumentManager\PathBuilder;
 use Sulu\Component\HttpCache\HandlerInvalidatePathInterface;
+use Sulu\Component\Webspace\CustomUrl;
+use Sulu\Component\Webspace\Environment;
+use Sulu\Component\Webspace\Manager\WebspaceManagerInterface;
+use Sulu\Component\Webspace\Portal;
 
 /**
  * Provides testcases for custom-url-manager.
  */
 class CustomUrlManagerTest extends \PHPUnit_Framework_TestCase
 {
-    private function getMapping()
+    /**
+     * @var DocumentManagerInterface
+     */
+    private $documentManager;
+
+    /**
+     * @var CustomUrlRepository
+     */
+    private $customUrlRepository;
+
+    /**
+     * @var MetadataFactoryInterface
+     */
+    private $metadataFactory;
+
+    /**
+     * @var PathBuilder
+     */
+    private $pathBuilder;
+
+    /**
+     * @var HandlerInvalidatePathInterface
+     */
+    private $cacheHandler;
+
+    /**
+     * @var WebspaceManagerInterface
+     */
+    private $webspaceManager;
+
+    /**
+     * @var string
+     */
+    private $environment;
+
+    /**
+     * @var PageDocument
+     */
+    private $targetDocument;
+
+    /**
+     * @var Metadata
+     */
+    private $metadata;
+
+    /**
+     * @var CustomUrlManager
+     */
+    private $manager;
+
+    protected function setUp()
     {
-        return [
-            'title' => ['property' => 'title'],
-            'published' => ['property' => 'published'],
-            'baseDomain' => ['property' => 'baseDomain'],
-            'domainParts' => ['property' => 'domainParts', 'type' => 'json_array'],
-            'targetDocument' => ['property' => 'targetDocument', 'type' => 'reference'],
-            'canonical' => ['property' => 'canonical'],
-            'redirect' => ['property' => 'redirect'],
-            'targetLocale' => ['property' => 'targetLocale'],
-        ];
+        $this->documentManager = $this->prophesize(DocumentManagerInterface::class);
+        $this->customUrlRepository = $this->prophesize(CustomUrlRepository::class);
+        $this->metadataFactory = $this->prophesize(MetadataFactoryInterface::class);
+        $this->pathBuilder = $this->prophesize(PathBuilder::class);
+        $this->cacheHandler = $this->prophesize(HandlerInvalidatePathInterface::class);
+        $this->webspaceManager = $this->prophesize(WebspaceManagerInterface::class);
+
+        $this->targetDocument = $this->prophesize(PageDocument::class)->reveal();
+        $this->metadata = $this->prophesize(Metadata::class);
+
+        $this->manager = new CustomUrlManager(
+            $this->documentManager->reveal(),
+            $this->customUrlRepository->reveal(),
+            $this->metadataFactory->reveal(),
+            $this->pathBuilder->reveal(),
+            $this->cacheHandler->reveal(),
+            $this->webspaceManager->reveal(),
+            $this->environment
+        );
     }
 
     public function testCreate()
     {
-        $documentManager = $this->prophesize(DocumentManagerInterface::class);
-        $customUrlRepository = $this->prophesize(CustomUrlRepository::class);
-        $metadataFactory = $this->prophesize(MetadataFactoryInterface::class);
-        $pathBuilder = $this->prophesize(PathBuilder::class);
-        $cacheHandler = $this->prophesize(HandlerInvalidatePathInterface::class);
-        $targetDocument = $this->prophesize(PageDocument::class)->reveal();
-
-        $metadata = $this->prophesize(Metadata::class);
-        $metadata->getFieldMappings()->willReturn($this->getMapping());
-        $metadataFactory->getMetadataForAlias('custom_url')->willReturn($metadata);
+        $this->metadata->getFieldMappings()->willReturn($this->getMapping());
+        $this->metadataFactory->getMetadataForAlias('custom_url')->willReturn($this->metadata);
 
         $testDocument = new CustomUrlDocument();
-        $documentManager->create('custom_url')->willReturn($testDocument);
-        $documentManager->persist(
+        $this->documentManager->create('custom_url')->willReturn($testDocument);
+        $this->documentManager->persist(
             $testDocument,
             'en',
             ['parent_path' => '/cmf/sulu_io/custom_urls/items', 'load_ghost_content' => true, 'auto_rename' => false]
         )->shouldBeCalledTimes(1);
-        $documentManager->publish($testDocument, 'en')->shouldBeCalled();
-        $documentManager->find('123-123-123', 'en', ['load_ghost_content' => true])
-            ->willReturn($targetDocument);
+        $this->documentManager->publish($testDocument, 'en')->shouldBeCalled();
+        $this->documentManager->find('123-123-123', 'en', ['load_ghost_content' => true])->willReturn(
+                $this->targetDocument
+            );
 
-        $pathBuilder->build(['%base%', 'sulu_io', '%custom_urls%', '%custom_urls_items%'])
+        $this->pathBuilder->build(['%base%', 'sulu_io', '%custom_urls%', '%custom_urls_items%'])
             ->willReturn('/cmf/sulu_io/custom_urls/items');
 
-        $manager = new CustomUrlManager(
-            $documentManager->reveal(),
-            $customUrlRepository->reveal(),
-            $metadataFactory->reveal(),
-            $pathBuilder->reveal(),
-            $cacheHandler->reveal()
-        );
-
-        $result = $manager->create(
+        $result = $this->manager->create(
             'sulu_io',
             [
                 'title' => 'Test',
@@ -100,7 +148,7 @@ class CustomUrlManagerTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('de', $result->getTargetLocale());
         $this->assertEquals('*.sulu.io', $result->getBaseDomain());
         $this->assertEquals(['prefix' => 'test-1', 'postfix' => ['test-1', 'test-2']], $result->getDomainParts());
-        $this->assertEquals($targetDocument, $result->getTargetDocument());
+        $this->assertEquals($this->targetDocument, $result->getTargetDocument());
         $this->assertTrue($result->isPublished());
         $this->assertTrue($result->isCanonical());
         $this->assertTrue($result->isRedirect());
@@ -108,148 +156,94 @@ class CustomUrlManagerTest extends \PHPUnit_Framework_TestCase
 
     public function testFindList()
     {
-        $documentManager = $this->prophesize(DocumentManagerInterface::class);
-        $customUrlRepository = $this->prophesize(CustomUrlRepository::class);
-        $metadataFactory = $this->prophesize(MetadataFactoryInterface::class);
-        $pathBuilder = $this->prophesize(PathBuilder::class);
-        $cacheHandler = $this->prophesize(HandlerInvalidatePathInterface::class);
-
-        $pathBuilder->build(['%base%', 'sulu_io', '%custom_urls%', '%custom_urls_items%'])
+        $this->pathBuilder->build(['%base%', 'sulu_io', '%custom_urls%', '%custom_urls_items%'])
             ->willReturn('/cmf/sulu_io/custom_urls/items');
 
-        $manager = new CustomUrlManager(
-            $documentManager->reveal(),
-            $customUrlRepository->reveal(),
-            $metadataFactory->reveal(),
-            $pathBuilder->reveal(),
-            $cacheHandler->reveal()
-        );
-
-        $customUrlRepository->findList('/cmf/sulu_io/custom_urls/items', 'de')
+        $this->customUrlRepository->findList('/cmf/sulu_io/custom_urls/items', 'de', ['*.sulu.io', 'sulu.io/*'])
             ->willReturn([['title' => 'Test-1'], ['title' => 'Test-2']]);
 
-        $result = $manager->findList('sulu_io', 'de');
+        $url1 = $this->prophesize(CustomUrl::class);
+        $url1->getUrl()->willReturn('*.sulu.io');
+
+        $url2 = $this->prophesize(CustomUrl::class);
+        $url2->getUrl()->willReturn('sulu.io/*');
+
+        $environment = $this->prophesize(Environment::class);
+        $environment->getCustomUrls()->willReturn([$url1->reveal(), $url2->reveal()]);
+
+        $portal = $this->prophesize(Portal::class);
+        $portal->getEnvironment($this->environment)->willReturn($environment->reveal());
+
+        $webspace = $this->prophesize(WebspaceManagerInterface::class);
+        $webspace->getPortals()->willReturn([$portal->reveal()]);
+
+        $this->webspaceManager->findWebspaceByKey('sulu_io')->willReturn($webspace->reveal());
+
+        $result = $this->manager->findList('sulu_io', 'de');
 
         $this->assertEquals([['title' => 'Test-1'], ['title' => 'Test-2']], $result);
     }
 
     public function testFindUrls()
     {
-        $documentManager = $this->prophesize(DocumentManagerInterface::class);
-        $customUrlRepository = $this->prophesize(CustomUrlRepository::class);
-        $metadataFactory = $this->prophesize(MetadataFactoryInterface::class);
-        $pathBuilder = $this->prophesize(PathBuilder::class);
-        $cacheHandler = $this->prophesize(HandlerInvalidatePathInterface::class);
-
-        $pathBuilder->build(['%base%', 'sulu_io', '%custom_urls%', '%custom_urls_items%'])
+        $this->pathBuilder->build(['%base%', 'sulu_io', '%custom_urls%', '%custom_urls_items%'])
             ->willReturn('/cmf/sulu_io/custom_urls/items');
 
-        $manager = new CustomUrlManager(
-            $documentManager->reveal(),
-            $customUrlRepository->reveal(),
-            $metadataFactory->reveal(),
-            $pathBuilder->reveal(),
-            $cacheHandler->reveal()
-        );
-
-        $customUrlRepository->findUrls('/cmf/sulu_io/custom_urls/items')
+        $this->customUrlRepository->findUrls('/cmf/sulu_io/custom_urls/items')
             ->willReturn(['1.sulu.lo', '1.sulu.lo/2']);
 
-        $result = $manager->findUrls('sulu_io');
+        $result = $this->manager->findUrls('sulu_io');
 
         $this->assertEquals(['1.sulu.lo', '1.sulu.lo/2'], $result);
     }
 
     public function testFind()
     {
-        $documentManager = $this->prophesize(DocumentManagerInterface::class);
-        $customUrlRepository = $this->prophesize(CustomUrlRepository::class);
-        $metadataFactory = $this->prophesize(MetadataFactoryInterface::class);
-        $pathBuilder = $this->prophesize(PathBuilder::class);
-        $cacheHandler = $this->prophesize(HandlerInvalidatePathInterface::class);
         $document = $this->prophesize(CustomUrlDocument::class);
 
-        $manager = new CustomUrlManager(
-            $documentManager->reveal(),
-            $customUrlRepository->reveal(),
-            $metadataFactory->reveal(),
-            $pathBuilder->reveal(),
-            $cacheHandler->reveal()
-        );
+        $this->documentManager->find('123-123-123', 'de', ['load_ghost_content' => true])
+            ->willReturn($document->reveal());
 
-        $documentManager->find('123-123-123', 'de', ['load_ghost_content' => true])->willReturn($document->reveal());
-
-        $result = $manager->find('123-123-123', 'de');
+        $result = $this->manager->find('123-123-123', 'de');
 
         $this->assertEquals($document->reveal(), $result);
     }
 
     public function testFindByUrl()
     {
-        $documentManager = $this->prophesize(DocumentManagerInterface::class);
-        $customUrlRepository = $this->prophesize(CustomUrlRepository::class);
-        $metadataFactory = $this->prophesize(MetadataFactoryInterface::class);
-        $pathBuilder = $this->prophesize(PathBuilder::class);
-        $cacheHandler = $this->prophesize(HandlerInvalidatePathInterface::class);
         $routeDocument = $this->prophesize(RouteDocument::class);
         $customUrlDocument = $this->prophesize(CustomUrlDocument::class);
 
         $routeDocument->getTargetDocument()->willReturn($customUrlDocument->reveal());
 
-        $pathBuilder->build(['%base%', 'sulu_io', '%custom_urls%', '%custom_urls_routes%'])
+        $this->pathBuilder->build(['%base%', 'sulu_io', '%custom_urls%', '%custom_urls_routes%'])
             ->willReturn('/cmf/sulu_io/custom_urls/routes');
 
-        $manager = new CustomUrlManager(
-            $documentManager->reveal(),
-            $customUrlRepository->reveal(),
-            $metadataFactory->reveal(),
-            $pathBuilder->reveal(),
-            $cacheHandler->reveal()
-        );
-
-        $documentManager->find('/cmf/sulu_io/custom_urls/routes/sulu.io/test', 'de', ['load_ghost_content' => true])
+        $this->documentManager->find('/cmf/sulu_io/custom_urls/routes/sulu.io/test', 'de', ['load_ghost_content' => true])
             ->willReturn($routeDocument->reveal());
 
-        $result = $manager->findByUrl('sulu.io/test', 'sulu_io', 'de');
+        $result = $this->manager->findByUrl('sulu.io/test', 'sulu_io', 'de');
 
         $this->assertEquals($customUrlDocument->reveal(), $result);
     }
 
     public function testFindRouteByUrl()
     {
-        $documentManager = $this->prophesize(DocumentManagerInterface::class);
-        $customUrlRepository = $this->prophesize(CustomUrlRepository::class);
-        $metadataFactory = $this->prophesize(MetadataFactoryInterface::class);
-        $pathBuilder = $this->prophesize(PathBuilder::class);
-        $cacheHandler = $this->prophesize(HandlerInvalidatePathInterface::class);
         $routeDocument = $this->prophesize(RouteDocument::class);
 
-        $pathBuilder->build(['%base%', 'sulu_io', '%custom_urls%', '%custom_urls_routes%'])
+        $this->pathBuilder->build(['%base%', 'sulu_io', '%custom_urls%', '%custom_urls_routes%'])
             ->willReturn('/cmf/sulu_io/custom_urls/routes');
 
-        $manager = new CustomUrlManager(
-            $documentManager->reveal(),
-            $customUrlRepository->reveal(),
-            $metadataFactory->reveal(),
-            $pathBuilder->reveal(),
-            $cacheHandler->reveal()
-        );
-
-        $documentManager->find('/cmf/sulu_io/custom_urls/routes/sulu.io/test', 'de', ['load_ghost_content' => true])
+        $this->documentManager->find('/cmf/sulu_io/custom_urls/routes/sulu.io/test', 'de', ['load_ghost_content' => true])
             ->willReturn($routeDocument->reveal());
 
-        $result = $manager->findRouteByUrl('sulu.io/test', 'sulu_io', 'de');
+        $result = $this->manager->findRouteByUrl('sulu.io/test', 'sulu_io', 'de');
 
         $this->assertEquals($routeDocument->reveal(), $result);
     }
 
     public function testUpdate()
     {
-        $documentManager = $this->prophesize(DocumentManagerInterface::class);
-        $customUrlRepository = $this->prophesize(CustomUrlRepository::class);
-        $metadataFactory = $this->prophesize(MetadataFactoryInterface::class);
-        $pathBuilder = $this->prophesize(PathBuilder::class);
-        $cacheHandler = $this->prophesize(HandlerInvalidatePathInterface::class);
         $document = $this->prophesize(CustomUrlDocument::class);
         $targetDocument = $this->prophesize(PageDocument::class);
 
@@ -266,23 +260,14 @@ class CustomUrlManagerTest extends \PHPUnit_Framework_TestCase
         $document->setDomainParts(['prefix' => 'test-1', 'postfix' => ['test-1', 'test-2']])->shouldBeCalled();
         $document->setTargetDocument($targetDocument->reveal())->shouldBeCalled();
 
-        $metadata = $this->prophesize(Metadata::class);
-        $metadata->getFieldMappings()->willReturn($this->getMapping());
-        $metadataFactory->getMetadataForAlias('custom_url')->willReturn($metadata);
+        $this->metadata->getFieldMappings()->willReturn($this->getMapping());
+        $this->metadataFactory->getMetadataForAlias('custom_url')->willReturn($this->metadata->reveal());
 
-        $manager = new CustomUrlManager(
-            $documentManager->reveal(),
-            $customUrlRepository->reveal(),
-            $metadataFactory->reveal(),
-            $pathBuilder->reveal(),
-            $cacheHandler->reveal()
-        );
-
-        $documentManager->find('312-312-312', 'en', ['load_ghost_content' => true])->willReturn($document->reveal());
-        $documentManager->find('123-123-123', 'en', ['load_ghost_content' => true])->willReturn(
+        $this->documentManager->find('312-312-312', 'en', ['load_ghost_content' => true])->willReturn($document->reveal());
+        $this->documentManager->find('123-123-123', 'en', ['load_ghost_content' => true])->willReturn(
             $targetDocument->reveal()
         );
-        $documentManager->persist(
+        $this->documentManager->persist(
             $document,
             'en',
             [
@@ -292,9 +277,9 @@ class CustomUrlManagerTest extends \PHPUnit_Framework_TestCase
                 'auto_name_locale' => 'en',
             ]
         )->shouldBeCalledTimes(1);
-        $documentManager->publish($document, 'en')->shouldBeCalledTimes(1);
+        $this->documentManager->publish($document, 'en')->shouldBeCalledTimes(1);
 
-        $result = $manager->save(
+        $result = $this->manager->save(
             '312-312-312',
             [
                 'title' => 'Test',
@@ -314,11 +299,6 @@ class CustomUrlManagerTest extends \PHPUnit_Framework_TestCase
 
     public function testUpdateItemExists()
     {
-        $documentManager = $this->prophesize(DocumentManagerInterface::class);
-        $customUrlRepository = $this->prophesize(CustomUrlRepository::class);
-        $metadataFactory = $this->prophesize(MetadataFactoryInterface::class);
-        $pathBuilder = $this->prophesize(PathBuilder::class);
-        $cacheHandler = $this->prophesize(HandlerInvalidatePathInterface::class);
         $document = $this->prophesize(CustomUrlDocument::class);
         $targetDocument = $this->prophesize(PageDocument::class);
 
@@ -335,23 +315,14 @@ class CustomUrlManagerTest extends \PHPUnit_Framework_TestCase
         $document->setDomainParts(['prefix' => 'test-1', 'postfix' => ['test-1', 'test-2']])->shouldBeCalled();
         $document->setTargetDocument($targetDocument->reveal())->shouldBeCalled();
 
-        $metadata = $this->prophesize(Metadata::class);
-        $metadata->getFieldMappings()->willReturn($this->getMapping());
-        $metadataFactory->getMetadataForAlias('custom_url')->willReturn($metadata);
+        $this->metadata->getFieldMappings()->willReturn($this->getMapping());
+        $this->metadataFactory->getMetadataForAlias('custom_url')->willReturn($this->metadata->reveal());
 
-        $manager = new CustomUrlManager(
-            $documentManager->reveal(),
-            $customUrlRepository->reveal(),
-            $metadataFactory->reveal(),
-            $pathBuilder->reveal(),
-            $cacheHandler->reveal()
-        );
-
-        $documentManager->find('312-312-312', 'en', ['load_ghost_content' => true])->willReturn($document->reveal());
-        $documentManager->find('123-123-123', 'en', ['load_ghost_content' => true])->willReturn(
+        $this->documentManager->find('312-312-312', 'en', ['load_ghost_content' => true])->willReturn($document->reveal());
+        $this->documentManager->find('123-123-123', 'en', ['load_ghost_content' => true])->willReturn(
             $targetDocument->reveal()
         );
-        $documentManager->persist(
+        $this->documentManager->persist(
             $document,
             'en',
             [
@@ -364,7 +335,7 @@ class CustomUrlManagerTest extends \PHPUnit_Framework_TestCase
 
         $this->setExpectedException(TitleAlreadyExistsException::class);
 
-        $manager->save(
+        $this->manager->save(
             '312-312-312',
             [
                 'title' => 'Test',
@@ -382,34 +353,16 @@ class CustomUrlManagerTest extends \PHPUnit_Framework_TestCase
 
     public function testDelete()
     {
-        $documentManager = $this->prophesize(DocumentManagerInterface::class);
-        $customUrlRepository = $this->prophesize(CustomUrlRepository::class);
-        $metadataFactory = $this->prophesize(MetadataFactoryInterface::class);
-        $pathBuilder = $this->prophesize(PathBuilder::class);
-        $cacheHandler = $this->prophesize(HandlerInvalidatePathInterface::class);
         $document = $this->prophesize(CustomUrlDocument::class);
 
-        $documentManager->find('123-123-123', null, ['load_ghost_content' => true])->willReturn($document->reveal());
-        $documentManager->remove($document->reveal())->shouldBeCalled();
+        $this->documentManager->find('123-123-123', null, ['load_ghost_content' => true])->willReturn($document->reveal());
+        $this->documentManager->remove($document->reveal())->shouldBeCalled();
 
-        $manager = new CustomUrlManager(
-            $documentManager->reveal(),
-            $customUrlRepository->reveal(),
-            $metadataFactory->reveal(),
-            $pathBuilder->reveal(),
-            $cacheHandler->reveal()
-        );
-
-        $manager->delete('123-123-123');
+        $this->manager->delete('123-123-123');
     }
 
     public function testDeleteRoute()
     {
-        $documentManager = $this->prophesize(DocumentManagerInterface::class);
-        $customUrlRepository = $this->prophesize(CustomUrlRepository::class);
-        $metadataFactory = $this->prophesize(MetadataFactoryInterface::class);
-        $pathBuilder = $this->prophesize(PathBuilder::class);
-        $cacheHandler = $this->prophesize(HandlerInvalidatePathInterface::class);
         $document = $this->prophesize(RouteDocument::class);
         $customUrlDocument = $this->prophesize(CustomUrlDocument::class);
 
@@ -417,18 +370,10 @@ class CustomUrlManagerTest extends \PHPUnit_Framework_TestCase
         $document->getPath()->willReturn('/cmf/sulu_io/custom_urls/routes/sulu.io/test');
         $document->getTargetDocument()->willReturn($customUrlDocument->reveal());
 
-        $documentManager->find('123-123-123')->willReturn($document->reveal());
-        $documentManager->remove($document->reveal())->shouldBeCalled();
+        $this->documentManager->find('123-123-123')->willReturn($document->reveal());
+        $this->documentManager->remove($document->reveal())->shouldBeCalled();
 
-        $manager = new CustomUrlManager(
-            $documentManager->reveal(),
-            $customUrlRepository->reveal(),
-            $metadataFactory->reveal(),
-            $pathBuilder->reveal(),
-            $cacheHandler->reveal()
-        );
-
-        $result = $manager->deleteRoute('sulu_io', '123-123-123');
+        $result = $this->manager->deleteRoute('sulu_io', '123-123-123');
         $this->assertEquals($document->reveal(), $result);
     }
 
@@ -436,11 +381,6 @@ class CustomUrlManagerTest extends \PHPUnit_Framework_TestCase
     {
         $this->setExpectedException(RouteNotRemovableException::class);
 
-        $documentManager = $this->prophesize(DocumentManagerInterface::class);
-        $customUrlRepository = $this->prophesize(CustomUrlRepository::class);
-        $metadataFactory = $this->prophesize(MetadataFactoryInterface::class);
-        $pathBuilder = $this->prophesize(PathBuilder::class);
-        $cacheHandler = $this->prophesize(HandlerInvalidatePathInterface::class);
         $document = $this->prophesize(RouteDocument::class);
         $customUrlDocument = $this->prophesize(CustomUrlDocument::class);
 
@@ -448,30 +388,17 @@ class CustomUrlManagerTest extends \PHPUnit_Framework_TestCase
         $document->getPath()->willReturn('/cmf/sulu_io/custom_urls/routes/sulu.io/test');
         $document->getTargetDocument()->willReturn($customUrlDocument->reveal());
 
-        $documentManager->find('123-123-123')->willReturn($document->reveal());
-        $documentManager->remove($document->reveal())->shouldNotBeCalled();
+        $this->documentManager->find('123-123-123')->willReturn($document->reveal());
+        $this->documentManager->remove($document->reveal())->shouldNotBeCalled();
 
-        $pathBuilder->build(['%base%', 'sulu_io', '%custom_urls%', '%custom_urls_routes%'])
+        $this->pathBuilder->build(['%base%', 'sulu_io', '%custom_urls%', '%custom_urls_routes%'])
             ->willReturn('/cmf/sulu_io/custom_urls/routes');
 
-        $manager = new CustomUrlManager(
-            $documentManager->reveal(),
-            $customUrlRepository->reveal(),
-            $metadataFactory->reveal(),
-            $pathBuilder->reveal(),
-            $cacheHandler->reveal()
-        );
-
-        $manager->deleteRoute('sulu_io', '123-123-123');
+        $this->manager->deleteRoute('sulu_io', '123-123-123');
     }
 
     public function testInvalidate()
     {
-        $documentManager = $this->prophesize(DocumentManagerInterface::class);
-        $customUrlRepository = $this->prophesize(CustomUrlRepository::class);
-        $metadataFactory = $this->prophesize(MetadataFactoryInterface::class);
-        $pathBuilder = $this->prophesize(PathBuilder::class);
-        $cacheHandler = $this->prophesize(HandlerInvalidatePathInterface::class);
         $document = $this->prophesize(CustomUrlDocument::class);
 
         $document->getRoutes()->willReturn(
@@ -481,43 +408,36 @@ class CustomUrlManagerTest extends \PHPUnit_Framework_TestCase
             ]
         );
 
-        $manager = new CustomUrlManager(
-            $documentManager->reveal(),
-            $customUrlRepository->reveal(),
-            $metadataFactory->reveal(),
-            $pathBuilder->reveal(),
-            $cacheHandler->reveal()
-        );
+        $this->manager->invalidate($document->reveal());
 
-        $manager->invalidate($document->reveal());
-
-        $cacheHandler->invalidatePath('sulu.io/en')->shouldBeCalled();
-        $cacheHandler->invalidatePath('sulu.io/de')->shouldBeCalled();
+        $this->cacheHandler->invalidatePath('sulu.io/en')->shouldBeCalled();
+        $this->cacheHandler->invalidatePath('sulu.io/de')->shouldBeCalled();
     }
 
     public function testInvalidateRoute()
     {
-        $documentManager = $this->prophesize(DocumentManagerInterface::class);
-        $customUrlRepository = $this->prophesize(CustomUrlRepository::class);
-        $metadataFactory = $this->prophesize(MetadataFactoryInterface::class);
-        $pathBuilder = $this->prophesize(PathBuilder::class);
-        $cacheHandler = $this->prophesize(HandlerInvalidatePathInterface::class);
         $document = $this->prophesize(RouteDocument::class);
 
         $document->getPath()->willReturn('/cmf/sulu_io/custom_urls/routes/sulu.io/en');
-        $pathBuilder->build(['%base%', 'sulu_io', '%custom_urls%', '%custom_urls_routes%'])
+        $this->pathBuilder->build(['%base%', 'sulu_io', '%custom_urls%', '%custom_urls_routes%'])
             ->willReturn('/cmf/sulu_io/custom_urls/routes');
 
-        $manager = new CustomUrlManager(
-            $documentManager->reveal(),
-            $customUrlRepository->reveal(),
-            $metadataFactory->reveal(),
-            $pathBuilder->reveal(),
-            $cacheHandler->reveal()
-        );
+        $this->manager->invalidateRoute('sulu_io', $document->reveal());
 
-        $manager->invalidateRoute('sulu_io', $document->reveal());
+        $this->cacheHandler->invalidatePath('sulu.io/en')->shouldBeCalled();
+    }
 
-        $cacheHandler->invalidatePath('sulu.io/en')->shouldBeCalled();
+    private function getMapping()
+    {
+        return [
+            'title' => ['property' => 'title'],
+            'published' => ['property' => 'published'],
+            'baseDomain' => ['property' => 'baseDomain'],
+            'domainParts' => ['property' => 'domainParts', 'type' => 'json_array'],
+            'targetDocument' => ['property' => 'targetDocument', 'type' => 'reference'],
+            'canonical' => ['property' => 'canonical'],
+            'redirect' => ['property' => 'redirect'],
+            'targetLocale' => ['property' => 'targetLocale'],
+        ];
     }
 }